+++ include/Tester/mpc/nonlinear_mpc_optimization_engine/kinematic_bicycle_model_op_en_cost_matrices_Y_max.hpp
@@ -7,20 +7,13 @@ namespace kinematic_bicycle_model_op_en_cost_matrices_Y_max {
 
 using namespace PythonNumpy;
 
-using SparseAvailable_cost_matrices_Y_max = SparseAvailable<
-    ColumnAvailable<false>,
-    ColumnAvailable<false>,
-    ColumnAvailable<false>,
-    ColumnAvailable<false>
->;
+using SparseAvailable_cost_matrices_Y_max =
+    SparseAvailable<ColumnAvailable<false>, ColumnAvailable<false>,
+                    ColumnAvailable<false>, ColumnAvailable<false>>;
 
-using type = SparseMatrix_Type<double, SparseAvailable_cost_matrices_Y_max>;
+using type = SparseMatrix_Type<float, SparseAvailable_cost_matrices_Y_max>;
 
-inline auto make(void) -> type {
-
-  return make_SparseMatrixEmpty<double, 4, 1>();
-
-}
+inline auto make(void) -> type { return make_SparseMatrixEmpty<float, 4, 1>(); }
 
 } // namespace kinematic_bicycle_model_op_en_cost_matrices_Y_max
